Split out constraint/arbitrary.rs
